requirements.txt
httpx==0.27.2
python-dotenv==1.0.1
python-telegram-bot==21.10
resend==2.6.0
